requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
Django==5.0.1
django-bootstrap4==24.1
django-jazzmin==2.6.0
django-shortuuidfield==0.1.3
django-taggit==5.0.1
pillow==10.2.0
psycopg2-binary==2.9.9
shortuuid==1.0.13
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.4
